+++ features/aave/components/SimulateSectionComponent.tsx
@@ -22,7 +22,7 @@ import { useTranslation } from 'next-i18next'
 import React, { useMemo } from 'react'
 import { Box } from 'theme-ui'
 
-function mapSimulation(simulation?: Simulation): string[] {
+export function mapSimulation(simulation?: Simulation): string[] {
   if (!simulation) return [formatCryptoBalance(zero), formatCryptoBalance(zero)]
   return [
     `${formatCryptoBalance(simulation.earningAfterFees)} ${simulation.token}`,
@@ -30,7 +30,7 @@ function mapSimulation(simulation?: Simulation): string[] {
   ]
 }
 
-const defaultYieldFields: FilterYieldFieldsType[] = ['7Days', '30Days', '90Days', '1Year']
+export const defaultYieldFields: FilterYieldFieldsType[] = ['7Days', '30Days', '90Days', '1Year']
 
 function transitionHasSwap(
   transition?: IMultiplyStrategy | IStrategy,
