+++ features/omni-kit/components/details-section/index.ts
@@ -16,3 +16,6 @@ export * from './parsers/useOmniCardDataTokensValue'
 export * from './parsers/useOmniCardDataLiquidationRatio'
 export * from './parsers/useOmniCardDataNetApy'
 export * from './parsers/useOmniCardDataVariableAnnualFee'
+export * from './parsers/useOmniCardDataApy'
+export * from './parsers/useOmniCardDataBreakEven'
+export * from './parsers/useOmniCardDataEntryFees'
