+++ features/omni-kit/protocols/aave-like/components/index.ts
@@ -3,3 +3,6 @@ export * from './AaveLikeDetailsSectionFooter'
 export * from './AaveLikeContentFooterBorrow'
 export * from './AaveLikeContentFooterMultiply'
 export * from './AaveLikeCostToBorrowContentCardModal'
+export * from './AaveLikeDetailsSectionContentManage'
+export * from './AaveLikeDetailsSectionContentYieldLoopOpen'
+export * from './AaveLikeYieldLoopRiskBanner'
